[OBEY-CAMPAIGN-8deed8b4] test(next): split next tests into focused files
